+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Adjustments to the stock app configuration, including the deletion of view_assets and the addition of file_locations. Review thoroughly before implementing.
 - Added stock configuration for the module class.
 - Added 404 no-route handlers to the stock Rona module.
+- The HTTP Response handler now processes several options: full, module, and attrs. Previously, this parameter was passed in as $data and was just passed to the view_assets config., which has now been fully replaced.
 - Improvements to RonaPHP Logger.
 
 ### v. 1.4.0